style(tests): host-tests clippy 风格批量清理 + 文档同步
对 host-tests 全套源码与测试做 clippy 风格批量修复,
同步更新审计脚本与编译问题登记文档:

host-tests clippy 修复 (按 lint 分类):
- single_range_in_vec_init: copy_user_exception_test.rs 改用 &[range] 切片
  + 文件级 allow 抑制误报
- needless_range_loop: elf_loader_racy_cell_test.rs / wasi_test.rs
  改用 enumerate().take() 迭代器
- assertions_on_constants: socket_max_sockets_test.rs /
  overlayfs_test.rs 移除无意义 assert!(true), 必要处函数级 allow
- result_unit_err: framekernel_bench.rs 多处函数级 allow
- len_without_is_empty: HostZapStore trait 新增默认 is_empty 实现
- vec_init_then_push: run_all 函数 allow (80+ 项基准)
- type_complexity: eash_cmd_parser_test.rs 引入 CmdCase 类型别名
- 其他: arc.rs / dedup.rs / hvfs.rs / fsx.rs / lib.rs /
  demand_paging_test.rs / exec_rollback_test.rs / execve_signal_state_test.rs /
  exfat_test.rs / ext2_test.rs / feature_attr_minimal_test.rs /
  i43_block_bridge_test.rs / kmalloc_irq_save_test.rs / mmap_pwm_test.rs /
  multi_ioapic_routing_test.rs / nic_probe_arch_neutral_test.rs /
  page_fault_vma_flags_test.rs / plan_b_inode_test.rs /
  td08_kernel_error_test.rs / td20_services_kernel_error_test.rs /
  td22_sigill_delivery_test.rs / td23_sigaltstack_syscall_test.rs /
  test_runner_init_test.rs 同类小幅清理

scripts/audit_volatile_access.py: 完善审计逻辑, 与新增 HeapEndRef/
FreeListHeadRef 同类 volatile 访问模式匹配.

docs/plan/test-compile-issues-2026-07-31.md: 同步本轮 9 错误 + 1 warning +
12 dead_code 修复结果, 登记 DECISION-022 实际为源码缺陷 (非工具链限制),
后续可推进至 QEMU 集成测试路径.

验证: cargo test -p host-tests 全部通过 +
双架构 release 0w0e + clippy 0w + 三审计通过.

+++ host-tests/tests/kmalloc_irq_save_test.rs
@@ -59,7 +59,7 @@ impl MockHeap {
         if size == 0 { return None; }
         let flags = acquire_lock(&self.lock);
         // 模拟分配逻辑
-        let ptr = size as usize;
+        let ptr = size;
         release_lock(&self.lock, &flags);
         Some(ptr)
     }
